move most overlapped methods from JsonIO to JsonUtilsJS
